Refactor: 결제 승인 취소 재시도 클래스 분리
