+++ src/main/java/com/dangdangsalon/DangdangSalonServerApplication.java
@@ -5,9 +5,11 @@ import java.util.TimeZone;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+import org.springframework.retry.annotation.EnableRetry;
 
 @SpringBootApplication
 @EnableJpaAuditing
+@EnableRetry
 public class DangdangSalonServerApplication {
 
 	public static void main(String[] args) {
