+++ src/main/java/com/dangdangsalon/domain/payment/service/PaymentApproveService.java
@@ -20,19 +20,8 @@ import com.dangdangsalon.domain.payment.repository.PaymentRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.MediaType;
-import org.springframework.retry.annotation.Backoff;
-import org.springframework.retry.annotation.Recover;
-import org.springframework.retry.annotation.Retryable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.reactive.function.client.WebClient;
-import org.springframework.web.reactive.function.client.WebClientResponseException;
-
-import java.time.LocalDateTime;
-import java.util.Base64;
-import java.util.Map;
 
 @Slf4j
 @Service
@@ -44,10 +33,7 @@ class PaymentApproveService {
     private final EstimateRequestRepository estimateRequestRepository;
     private final PaymentNotificationService paymentNotificationService;
     private final CouponRepository couponRepository;
-    private final WebClient webClient;
-
-    @Value("${toss.api.key}")
-    private String tossApiKey;
+    private final PaymentApproveRetryService retryService;
 
     @Value("${toss.api.approve-url}")
     private String tossApproveUrl;
@@ -59,11 +45,8 @@ class PaymentApproveService {
         Orders order = findOrderByTossOrderId(paymentApproveRequestDto.getOrderId());
         validateOrderAmount(order, paymentApproveRequestDto.getAmount());
 
-        PaymentApproveResponseDto paymentResponse = sendApprovalRequestToToss(
-                paymentApproveRequestDto,
-                tossApproveUrl,
-                idempotencyKey
-        );
+        PaymentApproveResponseDto paymentResponse = retryService.sendApprovalRequestToToss(
+                paymentApproveRequestDto, tossApproveUrl, idempotencyKey);
 
         Long couponId = paymentApproveRequestDto.getCouponId();
         Coupon coupon = null;
@@ -81,53 +64,6 @@ class PaymentApproveService {
         paymentNotificationService.sendNotificationToUser(order);
 
         return paymentResponse;
-    }
-
-    @Retryable(
-            retryFor = { WebClientResponseException.class },
-            maxAttempts = 3,
-            backoff = @Backoff(delay = 2000)
-    )
-    public PaymentApproveResponseDto sendApprovalRequestToToss(PaymentApproveRequestDto paymentApproveRequestDto, String url, String idempotencyKey) {
-        String authorizationHeader = "Basic " + Base64.getEncoder().encodeToString((tossApiKey + ":").getBytes());
-
-        try {
-            return webClient.post()
-                    .uri(url)
-                    .header(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE)
-                    .header(HttpHeaders.AUTHORIZATION, authorizationHeader)
-                    .header("Idempotency-Key", idempotencyKey)
-                    .bodyValue(Map.of(
-                            "paymentKey", paymentApproveRequestDto.getPaymentKey(),
-                            "orderId", paymentApproveRequestDto.getOrderId(),
-                            "amount", paymentApproveRequestDto.getAmount()
-                    ))
-                    .retrieve()
-                    .bodyToMono(PaymentApproveResponseDto.class)
-                    .block();
-        } catch (WebClientResponseException ex) {
-            log.error("Toss 결제 검증 API 호출 실패: {}", ex.getResponseBodyAsString());
-            throw new RuntimeException("결제 승인 중 오류가 발생했습니다.");
-        }
-    }
-
-    @Recover
-    public void recover(WebClientResponseException ex, PaymentApproveRequestDto requestDto) {
-        log.error("결제 승인 재시도 실패: 요청 ID={}, 오류={}", requestDto.getOrderId(), ex.getMessage());
-
-        Orders order = findOrderByTossOrderId(requestDto.getOrderId());
-
-        Payment rejectedPayment = Payment.builder()
-                .paymentKey(requestDto.getPaymentKey())
-                .totalAmount(0)
-                .paymentStatus(PaymentStatus.REJECTED)
-                .requestedAt(LocalDateTime.now())
-                .paymentMethod("UNKNOWN")
-                .orders(order)
-                .coupon(null)
-                .build();
-
-        paymentRepository.save(rejectedPayment);
     }
 
     private void validatePaymentAmount(long amount) {
