+++ src/main/java/com/dangdangsalon/domain/payment/service/PaymentCancelService.java
@@ -20,18 +20,8 @@ import com.dangdangsalon.domain.payment.repository.PaymentRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.MediaType;
-import org.springframework.retry.annotation.Backoff;
-import org.springframework.retry.annotation.Recover;
-import org.springframework.retry.annotation.Retryable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.reactive.function.client.WebClient;
-import org.springframework.web.reactive.function.client.WebClientResponseException;
-
-import java.util.Base64;
-import java.util.Map;
 
 @Slf4j
 @Service
@@ -42,10 +32,7 @@ class PaymentCancelService {
     private final EstimateRequestRepository estimateRequestRepository;
     private final CouponRepository couponRepository;
     private final OrdersRepository ordersRepository;
-    private final WebClient webClient;
-
-    @Value("${toss.api.key}")
-    private String tossApiKey;
+    private final PaymentCancelRetryService paymentCancelRetryService;
 
     @Value("${toss.api.cancel-url}")
     private String tossCancelUrl;
@@ -55,11 +42,8 @@ class PaymentCancelService {
         Payment payment = findPaymentByPaymentKey(paymentCancelRequestDto.getPaymentKey());
 
         String cancelUrl = tossCancelUrl.replace("{paymentKey}", paymentCancelRequestDto.getPaymentKey());
-        PaymentCancelResponseDto paymentCancelResponseDto = sendCancelRequestToToss(
-                paymentCancelRequestDto,
-                cancelUrl,
-                idempotencyKey
-        );
+
+        PaymentCancelResponseDto paymentCancelResponseDto = paymentCancelRetryService.sendCancelRequestToToss(paymentCancelRequestDto, cancelUrl, idempotencyKey);
 
         Long couponId = paymentCancelRequestDto.getCouponId();
 
@@ -77,42 +61,6 @@ class PaymentCancelService {
         updatePaymentAndRelatedEntities(payment);
 
         return buildPaymentCancelResponseDto(paymentCancelResponseDto);
-    }
-
-    @Retryable(
-            retryFor = { WebClientResponseException.class },
-            maxAttempts = 3,
-            backoff = @Backoff(delay = 2000)
-    )
-    public PaymentCancelResponseDto sendCancelRequestToToss(PaymentCancelRequestDto paymentCancelRequestDto, String url, String idempotencyKey) {
-
-        String authorizationHeader = "Basic " + Base64.getEncoder().encodeToString((tossApiKey + ":").getBytes());
-        try {
-            return webClient.post()
-                    .uri(url)
-                    .header(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE)
-                    .header(HttpHeaders.AUTHORIZATION, authorizationHeader)
-                    .header("Idempotency-Key", idempotencyKey)
-                    .bodyValue(Map.of(
-                            "cancelReason", paymentCancelRequestDto.getCancelReason()
-                    ))
-                    .retrieve()
-                    .bodyToMono(PaymentCancelResponseDto.class)
-                    .block();
-        } catch (WebClientResponseException ex) {
-            log.error("Toss 결제 취소 API 호출 실패: {}", ex.getResponseBodyAsString());
-            throw new RuntimeException("결제 취소 중 오류가 발생했습니다.");
-        }
-    }
-
-    @Recover
-    public void recover(WebClientResponseException ex, PaymentCancelRequestDto requestDto) {
-        log.error("결제 취소 재시도 실패: 요청 ID={}, 오류={}", requestDto.getPaymentKey(), ex.getMessage());
-
-        Payment payment = paymentRepository.findByPaymentKey(requestDto.getPaymentKey())
-                .orElseThrow(() -> new IllegalArgumentException("결제 정보를 찾을 수 없습니다."));
-
-        payment.updatePaymentStatus(PaymentStatus.REJECTED);
     }
 
     private Payment findPaymentByPaymentKey(String paymentKey) {
